Reduce signed/unsigned comparison warnings
The warnings generated by Clang. This is part 1 for some simple cases.

BUG=pdfium:29
R=tsepez@chromium.org

Review URL: https://codereview.chromium.org/1840483003 .

+++ core/fpdfapi/fpdf_parser/cpdf_array.cpp
@@ -47,7 +47,7 @@ const CPDF_Array* CPDF_Array::AsArray() const {
 
 CPDF_Object* CPDF_Array::Clone(FX_BOOL bDirect) const {
   CPDF_Array* pCopy = new CPDF_Array();
-  for (int i = 0; i < GetCount(); i++) {
+  for (size_t i = 0; i < GetCount(); i++) {
     CPDF_Object* value = m_Objects.GetAt(i);
     pCopy->m_Objects.Add(value->Clone(bDirect));
   }
